Support SO_BINDTODEVICE in ICMP sockets
Adds support for the SO_BINDTODEVICE socket option in ICMP sockets with an
accompanying packetimpact test to exercise use of this socket option.

Adds a unit test to exercise the NIC selection logic introduced by this change.
The remaining unit tests for ICMP sockets need to be added in a subsequent CL.
See https://gvisor.dev/issues/5623 for the list of remaining unit tests.

Adds a "timeout" field to PacketimpactTestInfo, necessary due to the long
runtime of the newly added packetimpact test.

Test:
  - //third_party/gvisor/pkg/tcpip/transport/icmp:icmp_x_test
  - //third_party/gvisor/test/packetimpact/tests
    - :generic_dgram_socket_send_recv_native_test
    - :generic_dgram_socket_send_recv_netstack_test

Fixes #5678
Fixes #4896
Updates #5623
Updates #5681
Updates #5763
Updates #5956
Updates #5966
Updates #5967

PiperOrigin-RevId: 362147728